+++ __tests__/integration/integration-test-app/.cdsrc.mtls-prod.json
@@ -1,0 +1,13 @@
+{
+    "requires": {
+        "toggles": true
+    },
+    "ord": {
+        "authentication": {
+            "cfMtls": true
+        },
+        "describedSystemVersion": {
+            "version": "0.0.1"
+        }
+    }
+}
